Increase usability of default email message
Adding the email message into a text block adds a button to copy the
plaintext within the block for users that contains no additional
formatting, making it easier to use. It also prevents the urls within
the message from being auto-linked by github which was the original
intention of using the code snipits in the default email message.